dbsp: Remove `SpineFueled`.
We're currently maintaining two separate spine implementations. I don't
think we have any use for the old spine that doesn't have an async
merger any more. This commit removes it ahead of spine API changes, so
we don't need to worry about fixing and testing the second unused
implementation.
